relocated netcdf specific to netcdf, collecton to geo_ts_repository_collection.py
