+++ shyft/tests/test_netcdf_geo_ts_repository.py
@@ -2,7 +2,7 @@
 from os import path
 from shyft import shyftdata_dir
 from shyft.repository.netcdf.geo_ts_repository import GeoTsRepository
-from shyft.repository.yaml_config import YamlContent
+from shyft.repository.netcdf.yaml_config import YamlContent
 from shyft.api import Calendar
 from shyft.api import YMDhms
 from shyft.api import UtcPeriod
